requirements.txt
python-dotenv==0.19.0
bs4==0.0.1
schedule==1.1.0
requests==2.25.1